+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalFormApprove.php
@@ -111,6 +111,18 @@ final class ProposalFormApprove extends ProposalFormBase {
       '#format' => 'filtered_html',
     ];
 
+    $form['consent'] = [
+      '#type' => 'checkbox',
+      '#title' => $this
+        ->t('Personal data storage consent'),
+      '#required' => TRUE,
+      '#default_value' => TRUE,
+      '#description' => $this->getAdminFormStateValue('consent_help'),
+      '#attributes' => [
+        'disabled' => 'disabled',
+      ],
+    ];
+
     $form['actions']['#type'] = 'actions';
     $form['actions']['submit'] = [
       '#type' => 'submit',